bug fix
-fixed the first time pop up message showing up if the user had previously signed in
-removed the finish() in settings activity in order to allow the first time log in pop up to finish its task
